re-organize the USB subsystem into modules
